+++ lib/smtp-connection-as-promised.d.ts
@@ -1,3 +1,9 @@
+/// <reference types="node" />
+/// <reference path="../typings/mime-node.d.ts" />
+/// <reference path="../typings/mail-composer.d.ts" />
+/// <reference path="../typings/nodemailer-shared.d.ts" />
+/// <reference path="../typings/smtp-connection.d.ts" />
+
 import { PromiseReadable } from 'promise-readable'
 
 import * as shared from 'nodemailer/lib/shared'